rework packaging, breaking the names and all other branches was not needed
 - revert to eos_data_distribution
