+++ debian/copyright
@@ -1,0 +1,41 @@
+Format: http://www.debian.org/doc/packaging-manuals/copyright-format/1.0/
+Upstream-Name: eos-data-distribution
+
+Files: *
+Copyright:
+ © 2016 Endless Mobile, Inc.
+License: LGPL-3+
+
+Files:
+ dbus-service.py
+ demo/simulate-dbus-consumer.py
+ tools/test_get_async.py
+ tools/test_publish_async_nfd.py
+Copyright:
+ © 2014–2016 Regents of the University of California
+License: LGPL-3+
+
+Files: setup.py
+Copyright:
+ © 2015 Collabora Ltd.
+ © 2016 Endless Mobile, Inc.
+License: LGPL-3+
+
+License: LGPL-3+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU Lesser General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
+ .
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU Lesser General Public License for more details.
+ .
+ You should have received a copy of the GNU Lesser General Public License
+ along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ A copy of the GNU Lesser General Public License is in the file COPYING.
+ .
+ On Debian systems the full text of the GNU Lesser General Public
+ License can be found in the `/usr/share/common-licenses/LGPL-3'
+ file.
