+++ debian/rules
@@ -1,0 +1,6 @@
+#! /usr/bin/make -f
+
+export PYBUILD_DISABLE=test
+#export DH_VERBOSE = 1
+%:
+	dh $@ --with python2 --buildsystem=pybuild
